+++ spectomate/converters/poetry_to_pip.py
@@ -1,0 +1,4 @@
+"""
+poetry_to_pip.py
+"""
+