+++ spectomate/schemas/poetry_schema.py
@@ -1,0 +1,4 @@
+"""
+poetry_schema.py
+"""
+